Crate google_remotebuildexecution2

This documentation was generated from Remote Build Execution crate version 1.0.9+20190702, where 20190702 is the exact revision of the remotebuildexecution:v2 schema built by the mako code generator v1.0.9.

Everything else about the Remote Build Execution v2 API can be found at the official documentation site. The original source code is on github.

Features

Handle the following Resources with ease from the central hub ...

action results

get and update

actions

execute

blobs

batch read, batch update, find missing and get tree

operations

wait execution

Other activities are ...

get capabilities

Not what you are looking for ? Find all other Google APIs in their Rust documentation index.

Structure of this Library

The API is structured into the following primary items:

Hub

a central object to maintain state and allow accessing all Activities

creates Method Builders which in turn allow access to individual Call Builders

Resources

primary types that you can apply Activities to

a collection of properties and Parts

Parts

a collection of properties

never directly used in Activities

Activities

operations to apply to Resources

All structures are marked with applicable traits to further categorize them and ease browsing.

Generally speaking, you can invoke Activities like this:

let r = hub.resource().activity(...).doit()

Or specifically ...

let r = hub.action_results().get(...).doit()
let r = hub.action_results().update(...).doit()

The resource() and activity(...) calls create builders. The second one dealing with Activities supports various methods to configure the impending operation (not shown here). It is made such that all required arguments have to be specified right away (i.e. (...)), whereas all optional ones can be build up as desired. The doit() method performs the actual communication with the server and returns the respective result.

Usage

Setting up your Project

To use this library, you would put the following lines into your Cargo.toml file:

[dependencies]
google-remotebuildexecution2 = "*"
# This project intentionally uses an old version of Hyper. See
# https://github.com/Byron/google-apis-rs/issues/173 for more
# information.
hyper = "^0.10"
hyper-rustls = "^0.6"
serde = "^1.0"
serde_json = "^1.0"
yup-oauth2 = "^1.0"

A complete example

extern crate hyper;
extern crate hyper_rustls;
extern crate yup_oauth2 as oauth2;
extern crate google_remotebuildexecution2 as remotebuildexecution2;
use remotebuildexecution2::{Result, Error};
use std::default::Default;
use oauth2::{Authenticator, DefaultAuthenticatorDelegate, ApplicationSecret, MemoryStorage};
use remotebuildexecution2::RemoteBuildExecution;
 
// Get an ApplicationSecret instance by some means. It contains the `client_id` and 
// `client_secret`, among other things.
let secret: ApplicationSecret = Default::default();
// Instantiate the authenticator. It will choose a suitable authentication flow for you, 
// unless you replace  `None` with the desired Flow.
// Provide your own `AuthenticatorDelegate` to adjust the way it operates and get feedback about 
// what's going on. You probably want to bring in your own `TokenStorage` to persist tokens and
// retrieve them from storage.
let auth = Authenticator::new(&secret, DefaultAuthenticatorDelegate,
                              hyper::Client::with_connector(hyper::net::HttpsConnector::new(hyper_rustls::TlsClient::new())),
                              <MemoryStorage as Default>::default(), None);
let mut hub = RemoteBuildExecution::new(hyper::Client::with_connector(hyper::net::HttpsConnector::new(hyper_rustls::TlsClient::new())), auth);
// You can configure optional parameters by calling the respective setters at will, and
// execute the final call using `doit()`.
// Values shown here are possibly random and not representative !
let result = hub.action_results().get("instanceName", "hash", "sizeBytes")
             .inline_stdout(false)
             .inline_stderr(true)
             .add_inline_output_files("erat")
             .doit();
 
match result {
    Err(e) => match e {
        // The Error enum provides details about what exactly happened.
        // You can also just use its `Debug`, `Display` or `Error` traits
         Error::HttpError(_)
        |Error::MissingAPIKey
        |Error::MissingToken(_)
        |Error::Cancelled
        |Error::UploadSizeLimitExceeded(_, _)
        |Error::Failure(_)
        |Error::BadRequest(_)
        |Error::FieldClash(_)
        |Error::JsonDecodeError(_, _) => println!("{}", e),
    },
    Ok(res) => println!("Success: {:?}", res),
}

Handling Errors

All errors produced by the system are provided either as Result enumeration as return value of the doit() methods, or handed as possibly intermediate results to either the Hub Delegate, or the Authenticator Delegate.

When delegates handle errors or intermediate values, they may have a chance to instruct the system to retry. This makes the system potentially resilient to all kinds of errors.

Uploads and Downloads

If a method supports downloads, the response body, which is part of the Result, should be read by you to obtain the media. If such a method also supports a Response Result, it will return that by default. You can see it as meta-data for the actual media. To trigger a media download, you will have to set up the builder by making this call: .param("alt", "media").

Methods supporting uploads can do so using up to 2 different protocols: simple and resumable. The distinctiveness of each is represented by customized doit(...) methods, which are then named upload(...) and upload_resumable(...) respectively.

Customization and Callbacks

You may alter the way an doit() method is called by providing a delegate to the Method Builder before making the final doit() call. Respective methods will be called to provide progress information, as well as determine whether the system should retry on failure.

The delegate trait is default-implemented, allowing you to customize it with minimal effort.

Optional Parts in Server-Requests

All structures provided by this library are made to be enocodable and decodable via json. Optionals are used to indicate that partial requests are responses are valid. Most optionals are are considered Parts which are identifiable by name, which will be sent to the server to indicate either the set parts of the request or the desired parts in the response.

Builder Arguments

Using method builders, you are able to prepare an action call by repeatedly calling it's methods. These will always take a single argument, for which the following statements are true.

PODs are handed by copy

strings are passed as &str

request values are moved

Arguments will always be copied or cloned into the builder, to make them independent of their original life times.

Structs

ActionExecuteCall
Execute an action remotely.

ActionMethods
A builder providing access to all methods supported on action resources. It is not used directly, but through the RemoteBuildExecution hub.

ActionResultGetCall
Retrieve a cached execution result.

ActionResultMethods
A builder providing access to all methods supported on actionResult resources. It is not used directly, but through the RemoteBuildExecution hub.

ActionResultUpdateCall
Upload a new execution result.

BlobBatchReadCall
Download many blobs at once.

BlobBatchUpdateCall
Upload many blobs at once.

BlobFindMissingCall
Determine if blobs are present in the CAS.

BlobGetTreeCall
Fetch the entire directory tree rooted at a node.

BlobMethods
A builder providing access to all methods supported on blob resources. It is not used directly, but through the RemoteBuildExecution hub.

BuildBazelRemoteExecutionV2PriorityCapabilitiesPriorityRange
Supported range of priorities, including boundaries.

BuildBazelRemoteExecutionV2SymlinkNode
A SymlinkNode represents a symbolic link.

BuildBazelRemoteExecutionV2Directory
A Directory represents a directory node in a file tree, containing zero or more children FileNodes, DirectoryNodes and SymlinkNodes. Each Node contains its name in the directory, either the digest of its content (either a file blob or a Directory proto) or a symlink target, as well as possibly some metadata about the file or directory.

BuildBazelRemoteExecutionV2FindMissingBlobsResponse
A response message for ContentAddressableStorage.FindMissingBlobs.

BuildBazelRemoteExecutionV2GetTreeResponse
A response message for ContentAddressableStorage.GetTree.

BuildBazelRemoteExecutionV2BatchReadBlobsRequest
A request message for ContentAddressableStorage.BatchReadBlobs.

BuildBazelRemoteExecutionV2PriorityCapabilities
Allowed values for priority in ResultsCachePolicy Used for querying both cache and execution valid priority ranges.

BuildBazelRemoteExecutionV2WaitExecutionRequest
A request message for WaitExecution.

BuildBazelRemoteExecutionV2BatchUpdateBlobsRequestRequest
A request corresponding to a single blob that the client wants to upload.

BuildBazelRemoteExecutionV2ActionResult
An ActionResult represents the result of an Action being run.

BuildBazelRemoteExecutionV2ServerCapabilities
A response message for Capabilities.GetCapabilities.

BuildBazelRemoteExecutionV2BatchReadBlobsResponse
A response message for ContentAddressableStorage.BatchReadBlobs.

BuildBazelRemoteExecutionV2DirectoryNode
A DirectoryNode represents a child of a Directory which is itself a Directory and its associated metadata.

BuildBazelRemoteExecutionV2OutputDirectory
An OutputDirectory is the output in an ActionResult corresponding to a directory's full contents rather than a single file.

BuildBazelRemoteExecutionV2ExecuteRequest
A request message for Execution.Execute.

BuildBazelRemoteExecutionV2ActionCacheUpdateCapabilities
Describes the server/instance capabilities for updating the action cache.

BuildBazelRemoteExecutionV2CacheCapabilities
Capabilities of the remote cache system.

BuildBazelRemoteExecutionV2FindMissingBlobsRequest
A request message for ContentAddressableStorage.FindMissingBlobs.

BuildBazelRemoteExecutionV2BatchReadBlobsResponseResponse
A response corresponding to a single blob that the client tried to download.

BuildBazelRemoteExecutionV2OutputSymlink
An OutputSymlink is similar to a Symlink, but it is used as an output in an ActionResult.

BuildBazelRemoteExecutionV2BatchUpdateBlobsResponseResponse
A response corresponding to a single blob that the client tried to upload.

BuildBazelRemoteExecutionV2FileNode
A FileNode represents a single file and associated metadata.

BuildBazelRemoteExecutionV2OutputFile
An OutputFile is similar to a FileNode, but it is used as an output in an ActionResult. It allows a full file path rather than only a name.

BuildBazelRemoteExecutionV2BatchUpdateBlobsResponse
A response message for ContentAddressableStorage.BatchUpdateBlobs.

BuildBazelRemoteExecutionV2Digest
A content digest. A digest for a given blob consists of the size of the blob and its hash. The hash algorithm to use is defined by the server, but servers SHOULD use SHA-256.

BuildBazelRemoteExecutionV2BatchUpdateBlobsRequest
A request message for ContentAddressableStorage.BatchUpdateBlobs.

BuildBazelRemoteExecutionV2ResultsCachePolicy
A ResultsCachePolicy is used for fine-grained control over how action outputs are stored in the CAS and Action Cache.

BuildBazelRemoteExecutionV2ExecutionPolicy
An ExecutionPolicy can be used to control the scheduling of the action.

BuildBazelRemoteExecutionV2ExecutionCapabilities
Capabilities of the remote execution system.

BuildBazelRemoteExecutionV2ExecutedActionMetadata
ExecutedActionMetadata contains details about a completed execution.

BuildBazelSemverSemVer
The full version of a given tool.

Chunk

ContentRange
Implements the Content-Range header, for serialization only

DefaultDelegate
A delegate with a conservative default implementation, which is used if no other delegate is set.

DummyNetworkStream

ErrorResponse
A utility to represent detailed errors we might see in case there are BadRequests. The latter happen if the sent parameters or request structures are unsound

GoogleLongrunningOperation
This resource represents a long-running operation that is the result of a network API call.

GoogleRpcStatus
The Status type defines a logical error model that is suitable for different programming environments, including REST APIs and RPC APIs. It is used by gRPC. Each Status message contains three pieces of data: error code, error message, and error details.

JsonServerError
A utility type which can decode a server response that indicates error

MethodGetCapabilityCall
GetCapabilities returns the server capabilities configuration of the remote endpoint. Only the capabilities of the services supported by the endpoint will be returned:

MethodInfo
Contains information about an API request.

MethodMethods
A builder providing access to all free methods, which are not associated with a particular resource. It is not used directly, but through the RemoteBuildExecution hub.

MultiPartReader
Provides a Read interface that converts multiple parts into the protocol identified by RFC2387. Note: This implementation is just as rich as it needs to be to perform uploads to google APIs, and might not be a fully-featured implementation.

OperationMethods
A builder providing access to all methods supported on operation resources. It is not used directly, but through the RemoteBuildExecution hub.

OperationWaitExecutionCall
Wait for an execution operation to complete. When the client initially makes the request, the server immediately responds with the current status of the execution. The server will leave the request stream open until the operation completes, and then respond with the completed operation. The server MAY choose to stream additional updates as execution progresses, such as to provide an update as to the state of the execution.

RangeResponseHeader

RemoteBuildExecution
Central instance to access all RemoteBuildExecution related resource activities

ResumableUploadHelper
A utility type to perform a resumable upload from start to end.

ServerError

ServerMessage

XUploadContentType
The X-Upload-Content-Type header.

Enums

Error

Scope
Identifies the an OAuth2 authorization scope. A scope is needed when requesting an authorization token.

Traits

CallBuilder
Identifies types which represent builders for a particular resource method

Delegate
A trait specifying functionality to help controlling any request performed by the API. The trait has a conservative default implementation.

Hub
Identifies the Hub. There is only one per library, this trait is supposed to make intended use more explicit. The hub allows to access all resource methods more easily.

MethodsBuilder
Identifies types for building methods of a particular resource type

NestedType
Identifies types which are only used by other types internally. They have no special meaning, this trait just marks them for completeness.

Part
Identifies types which are only used as part of other types, which usually are carrying the Resource trait.

ReadSeek
A utility to specify reader types which provide seeking capabilities too

RequestValue
Identifies types which are used in API requests.

Resource
Identifies types which can be inserted and deleted. Types with this trait are most commonly used by clients of this API.

ResponseResult
Identifies types which are used in API responses.

ToParts
A trait for all types that can convert themselves into a parts string

UnusedType
Identifies types which are not actually used by the API This might be a bug within the google API schema.

Functions

remove_json_null_values

Type Definitions

Result
A universal result type used as return for all calls.